+++ src/errors.js
@@ -1,3 +1,5 @@
+import axios from 'axios'
+
 import { toCamelCaseDeep } from './utils/case_converter'
 import { get } from 'lodash'
 
@@ -33,12 +35,13 @@ export class ServerErrorBase extends Error {
    * Wrap a raw axios error.
    *
    * @param {object} originalError Raw axios response.
-   * @param {axios} axios Axios instance used for request.
+   * @param {axios} axiosInstance Axios instance used for request.
    */
-  constructor (originalError, axios) {
+  constructor (originalError, axiosInstance) {
     super(originalError.message)
+
     this.originalError = originalError
-    this._axios = axios
+    this._axios = axiosInstance || axios.create()
   }
 
   /**
@@ -78,7 +81,7 @@ export class ServerErrorBase extends Error {
    * Use it to retry requests after 2FA.
    */
   retryRequest () {
-    let config = this.originalError.config
+    const config = this.originalError.config
     return this._axios(config)
   }
 }